requirements.txt
python-dotenv
packaging
googlesearch-python 
pycountry
fastapi
uvicorn
groq
phidata